Rename item command and query classes to include 'Adm' for administrative context
