+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/command/RegisterProductImageAdmCommand.java
@@ -1,0 +1,18 @@
+package ing.beribtur.facade.api.feature.adm.item.command;
+
+import ing.beribtur.accent.message.CommandRequest;
+import ing.beribtur.aggregate.item.entity.sdo.ProductImageCdo;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.util.Assert;
+
+@Getter
+@Setter
+public class RegisterProductImageAdmCommand extends CommandRequest<String> {
+    //
+    private ProductImageCdo productImageCdo;
+
+    public void validate() {
+        Assert.notNull(productImageCdo, "'productImageCdo' is required.");
+    }
+}
