+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/command/RegisterProductVariantAdmCommand.java
@@ -1,0 +1,19 @@
+package ing.beribtur.facade.api.feature.adm.item.command;
+
+import ing.beribtur.accent.message.CommandRequest;
+import ing.beribtur.aggregate.item.entity.sdo.ProductVariantCdo;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.util.Assert;
+
+@Getter
+@Setter
+public class RegisterProductVariantAdmCommand extends CommandRequest<String> {
+    //
+    private ProductVariantCdo productVariantCdo;
+
+    public void validate() {
+        //
+        Assert.notNull(productVariantCdo, "'productVariantCdo' is required.");
+    }
+}
