+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/query/FindProductAdmQuery.java
@@ -1,0 +1,17 @@
+package ing.beribtur.facade.api.feature.adm.item.query;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.util.Assert;
+
+@Getter
+@Setter
+public class FindProductAdmQuery {
+    //
+    private String productId;
+
+    public void validate() {
+        //
+        Assert.hasText(productId, "'productId' is required.");
+    }
+}
