+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/query/FindProductCategoryAdmQuery.java
@@ -1,0 +1,21 @@
+package ing.beribtur.facade.api.feature.adm.item.query;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.util.Assert;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class FindProductCategoryAdmQuery {
+    //
+    private String categoryId;
+
+    public void validate() {
+        //
+        Assert.hasText(categoryId, "'categoryId' is required.");
+    }
+}
