+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/query/FindProductImageAdmQuery.java
@@ -1,0 +1,17 @@
+package ing.beribtur.facade.api.feature.adm.item.query;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.util.Assert;
+
+@Getter
+@Setter
+public class FindProductImageAdmQuery {
+    //
+    private String imageId;
+
+    public void validate() {
+        //
+        Assert.hasText(imageId, "'imageId' is required.");
+    }
+}
