+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/query/FindProductVariantAdmQuery.java
@@ -1,0 +1,17 @@
+package ing.beribtur.facade.api.feature.adm.item.query;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.util.Assert;
+
+@Getter
+@Setter
+public class FindProductVariantAdmQuery {
+    //
+    private String variantId;
+
+    public void validate() {
+        //
+        Assert.hasText(variantId, "'variantId' is required.");
+    }
+}
