+++ beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/adm/item/rest/ItemAdmFlowFacade.java
@@ -1,0 +1,23 @@
+package ing.beribtur.facade.api.feature.adm.item.rest;
+
+import ing.beribtur.accent.message.CommandResponse;
+import ing.beribtur.facade.api.feature.adm.item.command.*;
+
+public interface ItemAdmFlowFacade {
+    //
+    CommandResponse<String> registerProductCategory(RegisterProductCategoryAdmCommand command);
+    CommandResponse<String> modifyProductCategory(ModifyProductCategoryAdmCommand command);
+    CommandResponse<String> removeProductCategory(RemoveProductCategoryAdmCommand command);
+
+    CommandResponse<String> registerProduct(RegisterProductAdmCommand command);
+    CommandResponse<String> modifyProduct(ModifyProductAdmCommand command);
+    CommandResponse<String> removeProduct(RemoveProductAdmCommand command);
+
+    CommandResponse<String> registerProductImage(RegisterProductImageAdmCommand command);
+    CommandResponse<String> modifyProductImage(ModifyProductImageAdmCommand command);
+    CommandResponse<String> removeProductImage(RemoveProductImageAdmCommand command);
+
+    CommandResponse<String> registerProductVariant(RegisterProductVariantAdmCommand command);
+    CommandResponse<String> modifyProductVariant(ModifyProductVariantAdmCommand command);
+    CommandResponse<String> removeProductVariant(RemoveProductVariantAdmCommand command);
+}
