--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/command/ModifyProductImageCommand.java
@@ -1,20 +1,0 @@
-package ing.beribtur.facade.api.feature.item.command;
-
-import ing.beribtur.accent.domain.NameValueList;
-import ing.beribtur.accent.message.CommandRequest;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-public class ModifyProductImageCommand extends CommandRequest<String> {
-    //
-    private String imageId;
-    private NameValueList nameValueList;
-
-    public void validate() {
-        Assert.hasText(imageId, "'imageId' is required.");
-        Assert.notNull(nameValueList, "'nameValueList' is required.");
-    }
-}
