--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/command/RegisterProductCategoryCommand.java
@@ -1,24 +1,0 @@
-package ing.beribtur.facade.api.feature.item.command;
-
-import ing.beribtur.accent.message.CommandRequest;
-import ing.beribtur.aggregate.item.entity.sdo.ProductCategoryCdo;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterProductCategoryCommand extends CommandRequest<String> {
-    //
-    private ProductCategoryCdo productCategoryCdo;
-
-    public void validate() {
-        //
-        Assert.notNull(productCategoryCdo, "'categoryCdo' is required.");
-    }
-
-}
