--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/command/RegisterProductCommand.java
@@ -1,19 +1,0 @@
-package ing.beribtur.facade.api.feature.item.command;
-
-import ing.beribtur.accent.message.CommandRequest;
-import ing.beribtur.aggregate.item.entity.sdo.ProductCdo;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-public class RegisterProductCommand extends CommandRequest<String> {
-    //
-    private ProductCdo productCdo;
-
-    public void validate() {
-        //
-        Assert.notNull(productCdo, "'productCdo' is required.");
-    }
-}
