--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/command/RemoveProductCategoryCommand.java
@@ -1,22 +1,0 @@
-package ing.beribtur.facade.api.feature.item.command;
-
-import ing.beribtur.accent.message.CommandRequest;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class RemoveProductCategoryCommand extends CommandRequest<String> {
-    //
-    private String categoryId;
-
-    public void validate() {
-        //
-        Assert.hasText(categoryId, "'categoryId' is required.");
-    }
-}
