--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/command/RemoveProductVariantCommand.java
@@ -1,18 +1,0 @@
-package ing.beribtur.facade.api.feature.item.command;
-
-import ing.beribtur.accent.message.CommandRequest;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-public class RemoveProductVariantCommand extends CommandRequest<String> {
-    //
-    private String variantId;
-
-    public void validate() {
-        //
-        Assert.hasText(variantId, "'variantId' is required.");
-    }
-}
