--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/query/FindProductCategoryQuery.java
@@ -1,21 +1,0 @@
-package ing.beribtur.facade.api.feature.item.query;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class FindProductCategoryQuery {
-    //
-    private String categoryId;
-
-    public void validate() {
-        //
-        Assert.hasText(categoryId, "'categoryId' is required.");
-    }
-}
