--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/query/FindProductImageQuery.java
@@ -1,17 +1,0 @@
-package ing.beribtur.facade.api.feature.item.query;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-public class FindProductImageQuery {
-    //
-    private String imageId;
-
-    public void validate() {
-        //
-        Assert.hasText(imageId, "'imageId' is required.");
-    }
-}
