--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/query/FindProductVariantQuery.java
@@ -1,17 +1,0 @@
-package ing.beribtur.facade.api.feature.item.query;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.util.Assert;
-
-@Getter
-@Setter
-public class FindProductVariantQuery {
-    //
-    private String variantId;
-
-    public void validate() {
-        //
-        Assert.hasText(variantId, "'variantId' is required.");
-    }
-}
