--- beribturing-facade/src/main/java/ing/beribtur/facade/api/feature/item/rest/ItemSeekResource.java
@@ -1,65 +1,0 @@
-package ing.beribtur.facade.api.feature.item.rest;
-
-import ing.beribtur.accent.message.QueryResponse;
-import ing.beribtur.aggregate.item.entity.Product;
-import ing.beribtur.aggregate.item.entity.ProductCategory;
-import ing.beribtur.aggregate.item.entity.ProductImage;
-import ing.beribtur.aggregate.item.entity.ProductVariant;
-import ing.beribtur.facade.api.feature.item.query.FindProductCategoryQuery;
-import ing.beribtur.facade.api.feature.item.query.FindProductImageQuery;
-import ing.beribtur.facade.api.feature.item.query.FindProductQuery;
-import ing.beribtur.facade.api.feature.item.query.FindProductVariantQuery;
-import ing.beribtur.feature.item.seek.ItemSeek;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/feature/item")
-public class ItemSeekResource implements ItemSeekFacade {
-    //
-    private final ItemSeek itemSeek;
-
-    @Override
-    @PostMapping("/find-product-category/query")
-    public QueryResponse<ProductCategory> findProductCategory(@RequestBody FindProductCategoryQuery query) {
-        //
-        query.validate();
-        String categoryId = query.getCategoryId();
-        ProductCategory response = itemSeek.findProductCategoryById(categoryId);
-        return new QueryResponse<ProductCategory>(response);
-    }
-
-    @Override
-    @PostMapping("/find-product/query")
-    public QueryResponse<Product> findProduct(@RequestBody FindProductQuery query) {
-        //
-        query.validate();
-        String productId = query.getProductId();
-        Product response = itemSeek.findProductById(productId);
-        return new QueryResponse<Product>(response);
-    }
-
-    @Override
-    @PostMapping("/find-product-image/query")
-    public QueryResponse<ProductImage> findProductImage(@RequestBody FindProductImageQuery query) {
-        //
-        query.validate();
-        String imageId = query.getImageId();
-        ProductImage response = itemSeek.findProductImageById(imageId);
-        return new QueryResponse<ProductImage>(response);
-    }
-
-    @Override
-    @PostMapping("/find-product-variant/query")
-    public QueryResponse<ProductVariant> findProductVariant(@RequestBody FindProductVariantQuery query) {
-        //
-        query.validate();
-        String variantId = query.getVariantId();
-        ProductVariant response = itemSeek.findProductVariantById(variantId);
-        return new QueryResponse<ProductVariant>(response);
-    }
-}
