--- beribturing-feature/src/main/java/ing/beribtur/feature/rental/action/check/RenterCheck.java
@@ -1,5 +1,0 @@
-package ing.beribtur.feature.rental.action.check;
-
-public class RenterCheck {
-    //
-}
