--- beribturing-feature/src/main/java/ing/beribtur/feature/rental/action/helper/RenterHelper.java
@@ -1,5 +1,0 @@
-package ing.beribtur.feature.rental.action.helper;
-
-public class RenterHelper {
-    //
-}
