--- beribturing-feature/src/main/java/ing/beribtur/feature/rental/flow/RenterFlow.java
@@ -1,10 +1,0 @@
-package ing.beribtur.feature.rental.flow;
-
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class RenterFlow {
-    //
-}
