--- beribturing-feature/src/main/java/ing/beribtur/feature/rental/seek/RenterSeek.java
@@ -1,10 +1,0 @@
-package ing.beribtur.feature.rental.seek;
-
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class RenterSeek {
-    //
-}
