--- beribturing-feature/src/main/java/ing/beribtur/feature/rental/store/RenterCustomStore.java
@@ -1,5 +1,0 @@
-package ing.beribtur.feature.rental.store;
-
-public interface RenterCustomStore {
-    //
-}
